+++ packages/akashic-cli-commons/src/CliConfig/CliConfigServe.ts
@@ -13,4 +13,6 @@ export interface CliConfigServe {
 	targetDirs?: string[];
 	openBrowser?: boolean;
 	preserveDisconnected?: boolean;
+	watch?: boolean;
+	experimentalOpen?: number;
 }