+++ packages/akashic-cli-lib-manage/src/update/cli.ts
@@ -1,11 +1,12 @@
 import * as fs from "fs";
 import * as path from "path";
-import * as commander from "commander";
+import { Command } from "commander";
 import { ConsoleLogger, CliConfigurationFile, CliConfigUpdate } from "@akashic/akashic-cli-commons";
 import { promiseUpdate } from "./update";
 
 var ver = JSON.parse(fs.readFileSync(path.resolve(__dirname, "..", "..", "package.json"), "utf8")).version;
 
+const commander = new Command();
 commander
 	.usage("[options] <moduleName...>")
 	.description("Update installed npm modules")
@@ -25,7 +26,8 @@ commander
 
 export function run(argv: string[]): void {
 	commander.parse(argv);
-	CliConfigurationFile.read(path.join(commander["cwd"] || process.cwd(), "akashic.config.js"), (error, configuration) => {
+	const options = commander.opts();
+	CliConfigurationFile.read(path.join(options["cwd"] || process.cwd(), "akashic.config.js"), (error, configuration) => {
 		if (error) {
 			console.error(error);
 			process.exit(1);
@@ -33,8 +35,8 @@ export function run(argv: string[]): void {
 
 		const conf = configuration.commandOptions.update || {};
 		cli({
-			cwd: commander.cwd ?? conf.cwd,
-			quiet: commander.quiet ?? conf.quiet,
+			cwd: options.cwd ?? conf.cwd,
+			quiet: options.quiet ?? conf.quiet,
 			args: commander.args ?? conf.args
 		});
 	});