+++ packages/akashic-cli-serve/src/server/index.ts
@@ -5,7 +5,7 @@ import * as http from "http";
 import * as express from "express";
 import * as bodyParser from "body-parser";
 import * as socketio from "socket.io";
-import * as commander from "commander";
+import { Command, OptionValues } from "commander";
 import * as chalk from "chalk";
 import * as open from "open";
 import { PlayManager, RunnerManager, setSystemLogger, getSystemLogger } from "@akashic/headless-driver";
@@ -17,8 +17,11 @@ import { serverGlobalConfig } from "./common/ServerGlobalConfig";
 import { createContentsRouter } from "./route/ContentsRoute";
 import { createHealthCheckRouter } from "./route/HealthCheckRoute";
 import { ServerContentLocator } from "./common/ServerContentLocator";
-import {  CliConfigurationFile, CliConfigServe, SERVICE_TYPES } from "@akashic/akashic-cli-commons";
+import { CliConfigurationFile } from "@akashic/akashic-cli-commons/lib/CliConfig/CliConfigurationFile";
+import { CliConfigServe } from "@akashic/akashic-cli-commons/lib/CliConfig/CliConfigServe";
+import { SERVICE_TYPES } from "@akashic/akashic-cli-commons/lib/ServiceType";
 import { PlayerIdStore } from "./domain/PlayerIdStore";
+import { ModTargetFlags, watchContent } from "./domain/GameConfigs";
 
 // 渡されたパラメータを全てstringに変換する
 // chalkを使用する場合、ログ出力時objectの中身を展開してくれないためstringに変換する必要がある
@@ -32,7 +35,7 @@ function convertToStrings(params: any[]): string[] {
 	});
 }
 
-async function cli(cliConfigParam: CliConfigServe) {
+async function cli(cliConfigParam: CliConfigServe, cmdOptions: OptionValues) {
 	if (cliConfigParam.port && isNaN(cliConfigParam.port)) {
 		console.error("Invalid --port option: " + cliConfigParam.port);
 		process.exit(1);
@@ -104,15 +107,15 @@ async function cli(cliConfigParam: CliConfigServe) {
 	}
 
 	let gameExternalFactory: () => any = () => undefined;
-	if (commander.serverExternalScript) {
+	if (cmdOptions.serverExternalScript) {
 		try {
-			gameExternalFactory = require(path.resolve(commander.serverExternalScript));
+			gameExternalFactory = require(path.resolve(cmdOptions.serverExternalScript));
 		} catch (e) {
-			getSystemLogger().error(`Failed to evaluating --server-external-script (${commander.serverExternalScript}): ${e}`);
+			getSystemLogger().error(`Failed to evaluating --server-external-script (${cmdOptions.serverExternalScript}): ${e}`);
 			process.exit(1);
 		}
 		if (typeof gameExternalFactory !== "function") {
-			getSystemLogger().error(`${commander.serverExternalScript}, given as --server-external-script, does not export a function`);
+			getSystemLogger().error(`${cmdOptions.serverExternalScript}, given as --server-external-script, does not export a function`);
 			process.exit(1);
 		}
 	}
@@ -132,6 +135,36 @@ async function cli(cliConfigParam: CliConfigServe) {
 	const app = express();
 	const httpServer = http.createServer(app);
 	const io = new socketio.Server(httpServer);
+
+	if (cliConfigParam.watch && cliConfigParam.targetDirs) {
+		console.log("Start watching contents");
+		for (let i = 0; i < cliConfigParam.targetDirs.length; i++) {
+			await watchContent(cliConfigParam.targetDirs[i], async (err: any, modTargetFlag: ModTargetFlags) => {
+				if (err) {
+					getSystemLogger().error(err.message);
+				}
+				if (modTargetFlag === ModTargetFlags.GameJson) {
+					console.log("Reflect changes of game.json");
+				}
+				// コンテンツに変更があったらplayを新規に作り直して再起動
+				const contentId = `${i}`;
+				const targetPlayIds: string[] = [];
+				// 対象のコンテンツIDしか分からないので先に対応するコンテンツのrunnerを全て停止させる
+				playStore.getPlayIdsFromContentId(contentId).forEach(playId => {
+					playStore.getRunners(playId).forEach(runner => {
+						runnerStore.stopRunner(runner.runnerId);
+					});
+					targetPlayIds.push(playId);
+				});
+				const playId = await playStore.createPlay(new ServerContentLocator({ contentId }));
+				const token = amflowManager.createPlayToken(playId, "", "", true, {});
+				const amflow = playStore.createAMFlow(playId);
+				await runnerStore.createAndStartRunner({ playId, isActive: true, token, amflow, contentId });
+				await playStore.resumePlayDuration(playId);
+				targetPlayIds.forEach(id => io.emit("playBroadcast", { playId: id, message: { type: "switchPlay", nextPlayId: playId } }));
+			});
+		}
+	}
 
 	app.set("views", path.join(__dirname, "..", "..", "views"));
 	app.set("view engine", "ejs");
@@ -193,7 +226,7 @@ async function cli(cliConfigParam: CliConfigServe) {
 		// サーバー起動のログに関してはSystemLoggerで使用していない色を使いたいので緑を選択
 		console.log(chalk.green(`Hosting ${targetDirs.join(", ")} on http://${serverGlobalConfig.hostname}:${serverGlobalConfig.port}`));
 		if (loadedPlaylogPlayId) {
-			console.log(`play(id: ${loadedPlaylogPlayId}) read playlog(path: ${path.join(process.cwd(), commander.debugPlaylog)}).`);
+			console.log(`play(id: ${loadedPlaylogPlayId}) read playlog(path: ${path.join(process.cwd(), cmdOptions.debugPlaylog)}).`);
 			url += `?playId=${loadedPlaylogPlayId}&mode=replay`;
 			console.log(`if access ${url}, you can show this play.`);
 		}
@@ -207,6 +240,7 @@ async function cli(cliConfigParam: CliConfigServe) {
 // TODOこのファイルを改名してcli.tsにする
 export async function run(argv: any): Promise<void> {
 	const ver = JSON.parse(fs.readFileSync(path.resolve(__dirname, "..", "..", "package.json"), "utf8")).version;
+	const commander = new Command();
 	commander
 		.version(ver)
 		.description("Development server for Akashic Engine to debug multiple-player games")
@@ -216,6 +250,7 @@ export async function run(argv: any): Promise<void> {
 		.option("-v, --verbose", `Display detailed information on console.`)
 		.option("-A, --no-auto-start", `Wait automatic startup of contents.`)
 		.option("-s, --target-service <service>", `Simulate the specified service. arguments: ${SERVICE_TYPES}`)
+		.option("-w, --watch", `Watch directories of asset`)
 		.option("--server-external-script <path>",
 			`Evaluate the given JS and assign it to Game#external of the server instances`)
 		.option("--debug-playlog <path>", `Specify path of playlog-json.`)
@@ -226,7 +261,8 @@ export async function run(argv: any): Promise<void> {
 		.option("--preserve-disconnected", "Disable auto closing for disconnected windows.")
 		.parse(argv);
 
-	CliConfigurationFile.read(path.join(commander["cwd"] || process.cwd(), "akashic.config.js"), async (error, configuration) => {
+	const options = commander.opts();
+	CliConfigurationFile.read(path.join(options["cwd"] || process.cwd(), "akashic.config.js"), async (error, configuration) => {
 		if (error) {
 			console.error(error);
 			process.exit(1);
@@ -234,19 +270,20 @@ export async function run(argv: any): Promise<void> {
 
 		const conf = configuration.commandOptions.serve || {};
 		const cliConfigParam: CliConfigServe = {
-			port: commander.port ?? conf.port,
-			hostname: commander.hostname ?? conf.hostname,
-			verbose: commander.verbose ?? conf.verbose,
-			autoStart: commander.autoStart ?? conf.autoStart,
-			targetService: commander.targetService ?? conf.targetService,
-			debugPlaylog: commander.debugPlaylog ?? conf.debugPlaylog,
-			debugUntrusted: commander.debugUntrusted ?? conf.debugUntrusted,
-			debugProxyAudio: commander.proxyAudio ?? conf.debugProxyAudio,
-			allowExternal: commander.allowExternal ?? conf.allowExternal,
+			port: options.port ?? conf.port,
+			hostname: options.hostname ?? conf.hostname,
+			verbose: options.verbose ?? conf.verbose,
+			autoStart: options.autoStart ?? conf.autoStart,
+			targetService: options.targetService ?? conf.targetService,
+			debugPlaylog: options.debugPlaylog ?? conf.debugPlaylog,
+			debugUntrusted: options.debugUntrusted ?? conf.debugUntrusted,
+			debugProxyAudio: options.proxyAudio ?? conf.debugProxyAudio,
+			allowExternal: options.allowExternal ?? conf.allowExternal,
 			targetDirs: commander.args.length > 0 ? commander.args : (conf.targetDirs ?? [process.cwd()]),
-			openBrowser: commander.openBrowser ?? conf.openBrowser,
-			preserveDisconnected: commander.preserveDisconnected ?? conf.preserveDisconnected
+			openBrowser: options.openBrowser ?? conf.openBrowser,
+			preserveDisconnected: options.preserveDisconnected ?? conf.preserveDisconnected,
+			watch: options.watch ?? conf.watch
 		};
-		await cli(cliConfigParam);
+		await cli(cliConfigParam, options);
 	});
 }